Add files and fixtures for Cypress
