--- cypress/fixtures/artData.json
@@ -1,1 +1,0 @@
-{}
