+++ cypress/fixtures/artData1.json
@@ -1,0 +1,63 @@
+{
+  "objectID": 218013,
+  "isHighlight": false,
+  "accessionNumber": "09.68.75",
+  "accessionYear": "1909",
+  "isPublicDomain": true,
+  "primaryImage": "https://images.metmuseum.org/CRDImages/es/original/AR780.jpg",
+  "primaryImageSmall": "https://images.metmuseum.org/CRDImages/es/web-large/AR780.jpg",
+  "additionalImages": [
+    "https://images.metmuseum.org/CRDImages/es/original/GT93a.jpg",
+    "https://images.metmuseum.org/CRDImages/es/original/GT93b.jpg",
+    "https://images.metmuseum.org/CRDImages/es/original/GT93e.jpg"
+  ],
+  "constituents": null,
+  "department": "European Sculpture and Decorative Arts",
+  "objectName": "Border",
+  "title": "Border",
+  "culture": "",
+  "period": "",
+  "dynasty": "",
+  "reign": "",
+  "portfolio": "",
+  "artistRole": "",
+  "artistPrefix": "",
+  "artistDisplayName": "",
+  "artistDisplayBio": "",
+  "artistSuffix": "",
+  "artistAlphaSort": "",
+  "artistNationality": "",
+  "artistBeginDate": "",
+  "artistEndDate": "",
+  "artistGender": "",
+  "artistWikidata_URL": "",
+  "artistULAN_URL": "",
+  "objectDate": "ca. 1600",
+  "objectBeginDate": 1595,
+  "objectEndDate": 1605,
+  "medium": "Bobbin lace",
+  "dimensions": "L. 60 x W. 5 inches (152.4 x 12.7 cm)",
+  "measurements": null,
+  "creditLine": "Purchase by subscription, 1909",
+  "geographyType": "",
+  "city": "",
+  "state": "",
+  "county": "",
+  "country": "",
+  "region": "",
+  "subregion": "",
+  "locale": "",
+  "locus": "",
+  "excavation": "",
+  "river": "",
+  "classification": "Textiles-Laces",
+  "rightsAndReproduction": "",
+  "linkResource": "",
+  "metadataDate": "2020-03-02T21:50:01.377Z",
+  "repository": "Metropolitan Museum of Art, New York, NY",
+  "objectURL": "https://www.metmuseum.org/art/collection/search/218013",
+  "tags": null,
+  "objectWikidata_URL": "",
+  "isTimelineWork": false,
+  "GalleryNumber": ""
+}
