+++ cypress/fixtures/artData2.json
@@ -1,0 +1,87 @@
+{
+    "objectID": 203013,
+    "isHighlight": false,
+    "accessionNumber": "64.93.2",
+    "accessionYear": "1964",
+    "isPublicDomain": true,
+    "primaryImage": "https://images.metmuseum.org/CRDImages/es/original/DP104979.jpg",
+    "primaryImageSmall": "https://images.metmuseum.org/CRDImages/es/web-large/DP104979.jpg",
+    "additionalImages": [],
+    "constituents": [
+        {
+            "constituentID": 35652,
+            "role": "Artist",
+            "name": "Probably by Jean-Pierre Defrance",
+            "constituentULAN_URL": "http://vocab.getty.edu/page/ulan/500372490",
+            "constituentWikidata_URL": "",
+            "gender": ""
+        }
+    ],
+    "department": "European Sculpture and Decorative Arts",
+    "objectName": "Statue",
+    "title": "Fire",
+    "culture": "",
+    "period": "",
+    "dynasty": "",
+    "reign": "",
+    "portfolio": "",
+    "artistRole": "Artist",
+    "artistPrefix": "Probably by",
+    "artistDisplayName": "Jean-Pierre Defrance",
+    "artistDisplayBio": "1694–1768",
+    "artistSuffix": "",
+    "artistAlphaSort": "DEFRANCE JEAN-PIERRE",
+    "artistNationality": "",
+    "artistBeginDate": "1694",
+    "artistEndDate": "1768",
+    "artistGender": "",
+    "artistWikidata_URL": "",
+    "artistULAN_URL": "http://vocab.getty.edu/page/ulan/500372490",
+    "objectDate": "ca. 1750–60",
+    "objectBeginDate": 1745,
+    "objectEndDate": 1765,
+    "medium": "Limestone",
+    "dimensions": "Height: 7 ft. 3 3/4 in. (223 cm)",
+    "measurements": [
+        {
+            "elementName": "Height",
+            "elementDescription": null,
+            "elementMeasurements": {
+                "Height": 223
+            }
+        }
+    ],
+    "creditLine": "Purchase, Charles Ulrick and Josephine Bay Foundation Inc. Gift, 1964",
+    "geographyType": "",
+    "city": "",
+    "state": "",
+    "county": "",
+    "country": "",
+    "region": "",
+    "subregion": "",
+    "locale": "",
+    "locus": "",
+    "excavation": "",
+    "river": "",
+    "classification": "Sculpture",
+    "rightsAndReproduction": "",
+    "linkResource": "",
+    "metadataDate": "2020-09-16T18:35:19.457Z",
+    "repository": "Metropolitan Museum of Art, New York, NY",
+    "objectURL": "https://www.metmuseum.org/art/collection/search/203013",
+    "tags": [
+        {
+            "term": "Women",
+            "AAT_URL": "http://vocab.getty.edu/page/aat/300025943",
+            "Wikidata_URL": "https://www.wikidata.org/wiki/Q467"
+        },
+        {
+            "term": "Putti",
+            "AAT_URL": "http://vocab.getty.edu/page/aat/300250465",
+            "Wikidata_URL": "https://www.wikidata.org/wiki/Q284865"
+        }
+    ],
+    "objectWikidata_URL": "",
+    "isTimelineWork": true,
+    "GalleryNumber": "548"
+}
