+++ cypress/fixtures/artIDs.json
@@ -1,0 +1,10 @@
+[
+    208013,
+    203013,
+    10464,
+    10482,
+    11130,
+    470328,
+    11417,
+    11145
+]
