+++ src/components/App/App.scss
@@ -16,6 +16,7 @@
 
 }
 
+// MEDIA _ TABLET
 @media only screen and (max-width: 500px) {
   .App {
     background-image: url('../../rococo_mobile_art.jpg');
