+++ src/components/Artwork/Artwork.scss
@@ -9,6 +9,38 @@ img {
   margin: auto;
 }
 
+.div1 { 
+  grid-area: 1 / 6 / 4 / 9; 
+  align-self: end;  
+}
+.div2 { 
+  grid-area: 1 / 9 / 5 / 11;
+  align-self: end;  
+}
+.div3 { 
+  grid-area: 5 / 9 / 9 / 11; 
+  align-self: start; 
+}
+.div4 { 
+  grid-area: 5 / 4 / 9 / 6; 
+  align-self: start;
+}
+.div5 { 
+  grid-area: 4 / 2 / 6 / 4; 
+  align-self: start; 
+}
+.div6 { 
+  grid-area: 1 / 1 / 4 / 3; 
+  align-self: end; 
+  z-index: 0;
+}
+.div7 { grid-area: 2 / 3 / 4 / 6; align-self: end; z-index: 0;}
+.div8 { grid-area: 4 / 6 / 7 / 9; align-self: start;}
+.div9 { grid-area: 6 / 1 / 9 / 4; align-self: start;}
+.div10 { grid-area: 7 / 4 / 9 / 6; align-self: start; }
+.div11 { grid-area: 4 / 5 / 4 / 6; align-self: end; }
+.div12 { grid-area: 1 / 3 / 2 / 4; align-self: end; }
+.div13 { grid-area: 5 / 1 / 6 / 2; }
 // .div1 { grid-area: 1 / 6 / 7 / 9;
 //   align-self: end; }
 // .div2 { grid-area: 2 / 3 / 5 / 6;
@@ -25,19 +57,6 @@ img {
 // .div8 { grid-area: 5 / 1 / 7 / 4; }
 // .div9 { grid-area: 7 / 2 / 9 / 4; }
 
-.div1 { grid-area: 1 / 6 / 4 / 9; align-self: end;  }
-.div2 { grid-area: 1 / 9 / 5 / 11; align-self: end;  }
-.div3 { grid-area: 5 / 9 / 9 / 11; align-self: start; }
-.div4 { grid-area: 5 / 4 / 9 / 6; align-self: start;}
-.div5 { grid-area: 4 / 2 / 6 / 4; align-self: start; }
-.div6 { grid-area: 1 / 1 / 4 / 3; align-self: end; z-index: 0;}
-.div7 { grid-area: 2 / 3 / 4 / 6; align-self: end; z-index: 0;}
-.div8 { grid-area: 4 / 6 / 7 / 9; align-self: start;}
-.div9 { grid-area: 6 / 1 / 9 / 4; align-self: start;}
-.div10 { grid-area: 7 / 4 / 9 / 6; align-self: start; }
-.div11 { grid-area: 4 / 5 / 4 / 6; align-self: end; }
-.div12 { grid-area: 1 / 3 / 2 / 4; align-self: end; }
-.div13 { grid-area: 5 / 1 / 6 / 2; }
 // .div0 { grid-area: 1 / 2 / 1 / 4; }
 // .div1 { grid-area: 1 / 7 / 1 / 8; }
 // .div2 { grid-area: 2 / 1 / 3 / 4; }
