+++ src/components/Wall/Wall.scss
@@ -5,7 +5,7 @@
   grid-column-gap: 8px;
   grid-row-gap: 8px;
   width: 80%;
-  max-height: 700px;
+  max-height: 600px;
   margin: 15px auto;
   margin-top: -5em;
 }
